Refactor Camera Activity and associated classes as a package
