+++ app/src/main/res/layout/cameraactivity.xml
@@ -21,7 +21,8 @@
         android:id="@+id/CameraImageView"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:background="@color/cardview_dark_background"
+        android:background="@color/cardview_shadow_end_color"
+        android:rotation="90"
         android:visibility="gone"
         app:srcCompat="@color/cardview_dark_background" />
